// src/API.ts
export type CreatePostInput = {
  id?: string | null;
  title: string;
  content: string;
  image?: string | null;
};
// ...
export type Post = {
  __typename: 'Post';
  id: string;
  title: string;
  content: string;
  image?: string | null;
  createdAt: string;
  updatedAt: string;
  owner?: string | null;
  comments?: ModelCommentConnection | null;
  votes?: ModelVoteConnection | null;
};

export type ModelCommentConnection = {
  __typename: 'ModelCommentConnection';
  items?: Array<Comment | null> | null;
  nextToken?: string | null;
};

export type Comment = {
  __typename: 'Comment';
  id: string;
  postID: string;
  content: string;
  createdAt: string;
  updatedAt: string;
  post?: Post | null;
  owner?: string | null;
};

export type ModelVoteConnection = {
  __typename: 'ModelVoteConnection';
  items?: Array<Vote | null> | null;
  nextToken?: string | null;
};

export type Vote = {
  __typename: 'Vote';
  id: string;
  vote: string;
  postID: string;
  createdAt: string;
  updatedAt: string;
  post?: Post | null;
  owner?: string | null;
};
// ...
export type CreatePostMutation = {
  createPost?: {
    __typename: 'Post';
    id: string;
    title: string;
    content: string;
    image?: string | null;
    createdAt: string;
    updatedAt: string;
    owner?: string | null;
    comments?: {
      __typename: 'ModelCommentConnection';
      items?: Array<{
        __typename: 'Comment';
        id: string;
        postID: string;
        content: string;
        createdAt: string;
        updatedAt: string;
        post?: {
          __typename: 'Post';
          id: string;
          title: string;
          content: string;
          image?: string | null;
          createdAt: string;
          updatedAt: string;
          owner?: string | null;
        } | null;
        owner?: string | null;
      } | null> | null;
      nextToken?: string | null;
    } | null;
    votes?: {
      __typename: 'ModelVoteConnection';
      items?: Array<{
        __typename: 'Vote';
        id: string;
        vote: string;
        postID: string;
        createdAt: string;
        updatedAt: string;
        post?: {
          __typename: 'Post';
          id: string;
          title: string;
          content: string;
          image?: string | null;
          createdAt: string;
          updatedAt: string;
          owner?: string | null;
        } | null;
        owner?: string | null;
      } | null> | null;
      nextToken?: string | null;
    } | null;
  } | null;
};